refactor(threads): split transcript contract modules
